refactor: reuse card dimension helper
