+++ frontend/app/app/(app)/vertical-image-scroll/index.tsx
@@ -7,6 +7,7 @@ import { TranslationKeys } from '@/locales/keys';
 import useSetPageTitle from '@/hooks/useSetPageTitle';
 import { useSelector } from 'react-redux';
 import { RootState } from '@/redux/reducer';
+import CardDimensionHelper from '@/helper/CardDimensionHelper';
 
 const IMAGE_URLS = Array.from({ length: 20 }).map(
   (_, i) => `https://picsum.photos/id/${i + 10}/600/600`
@@ -30,30 +31,10 @@ const VerticalImageScroll = () => {
     return () => sub?.remove();
   }, []);
 
-  const getCardDimension = () => {
-    if (screenWidth < 1110 && screenWidth > 960) return 300;
-    else if (screenWidth < 840 && screenWidth > 750) return 350;
-    else if (screenWidth < 750 && screenWidth > 710) return 330;
-    else if (screenWidth < 709 && screenWidth > 650) return 300;
-    else if (screenWidth > 570) return 260;
-    else if (screenWidth > 530) return 240;
-    else if (screenWidth > 500) return 220;
-    else if (screenWidth > 450) return 210;
-    else if (screenWidth > 380) return 180;
-    else if (screenWidth > 360) return 170;
-    else if (screenWidth > 340) return 160;
-    else if (screenWidth > 320) return 150;
-    else if (screenWidth > 300) return 140;
-    else if (screenWidth > 280) return 130;
-    else return 120;
-  };
-
-  const getCardWidth = () => {
-    const offset = screenWidth < 500 ? 10 : screenWidth < 900 ? 25 : 35;
-    return screenWidth / numColumns - offset;
-  };
-
-  const size = amountColumnsForcard === 0 ? getCardDimension() : getCardWidth();
+  const size =
+    amountColumnsForcard === 0
+      ? CardDimensionHelper.getCardDimension(screenWidth)
+      : CardDimensionHelper.getCardWidth(screenWidth, numColumns);
 
   const [images, setImages] = useState<string[]>([]);
   const flatListRef = useRef<FlatList<string>>(null);
